Used SQLServer isntead Inmemory DB for now, Updated the table Schema to handle decimal values
Updated the model and DB Table to handle the dacimal values for tax numbers. Corrected the error msg and GetTaxRate endpoint query parameters.

+++ TaxService/Models/MunicipalityTaxRecord.cs
@@ -9,6 +9,6 @@ namespace MunicipalityTaxService.Models
     public DateTime ScheduleEnd { get; set; }
     public string MunicipalityName { get; set; }
     public TimePeriod Period { get; set; }
-    public double TaxRate { get; set; }
+    public Decimal TaxRate { get; set; }
     }
 }
